Deleting one tab to unblock loc
Hi @mihart,
We have a localizability issue for the sentence "Use the Query Editor to add an Index Column to your dataset." because it starts with two tabs, MD transformer treats that sentence as code and blocks it from being localizable. This can be fixed by deleting one tab.
Could you please help to signoff on this PR if you agree?