Request to fix broken bookmark link "Placeholder"
Hi @guyinacube,
We found from triaging loc bug that the bookmark link above for "Placeholder" does not work, because there is no heading for "Placeholder" in this article. Please help to fix as intended. I added comment tag only to indicate where the issue is, so you may delete the tag and update this PR or create separate PR to fix and I can close this one.