Removing Free Section
The license implementation has not reached production yet so I'm removing the section. However, we will add it back. Putting the section in this comment so we can put it back when the licensing impl. reaches prod.

//start//

## Allowing Free users access to reports

Free users can view reports that are embedded with the Power BI web part for SharePoint Online. You provide access to free users in the same way you do Pro users as described in [Granting access to reports](#granting-access-to-reports) above. The workspace, where the report is located, also has to be backed by Power BI Premium capacity. 

For example, if you have a report, in an app workspace, you would have to assign the app workspace to a Power BI Premium capacity. You would also need to add the Free user to the list of members of that app workspace.

//end//