Clarified the meaning of "latest data" as it was misleading
The data that you see in Power BI isn't the latest data. Tabular Models are generally processed periodically rather than themselves using Direct Query tables. The data that you see in Power BI is the latest data that is available to the Tabular Model, which will usually be the last data processed, not the "latest data".