C35225: Possible bad formatted bold and italic syntax
Hello, @maggies,
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.  
"There are extra asterisks at the end of the string and it is not clear whether the writer wants to make bold type for the phrase "this year sale" or leave it all in italics"

Please review and merge the proposed file change to fix to target versions. If you make related fix in another PR  then share your PR number so we can confirm and close this PR.
Many thanks in advance.