M34005: Possible Error: Direct Query on Live Connect
Hello @mgblythe,
Translator has reported possible source content issue.
Description:
The part in brackets seems to be faulty: Is it possible that the author means "DirectQuery or Live Connect" as these are two different methods of connecting data (http://radacad.com/directquery-live-connection-or-import-data-tough-decision)? Later on in the text ("Follow this guidance whether you're refreshing reports in import mode, or you're accessing data or data models by using DirectQuery or Live Connect.") "or" is used.
Please review and merge the suggested proposed file change to avoid this error. If you make related fix in another PR, then share your PR number, so we can confirm and close this PR.
Many thanks in advance.