Update Connect Excel to Power BI online
Based on the test on the latest version of Power BI Online and Excel 2016, when use the option "Connect, manage and view Excel in Power BI", Power BI Online will automatically create a dateset for the connected Excel workbook. But the current docs said "no dataset is created in Power BI". This is incorrect description, so i want to update it.