Update the description about upload workbook to Power BI
Based on the test on the latest version of Power BI Online and Excel 2016, when use the option "Upload your workbook to Power B", Power BI Online will automatically create a dataset for the connected Excel workbook. But the current docs said "no dataset is created in Power BI". This is incorrect description, so i want to update it.