Update service-premium-what-is.md
Added bars at the bottom of the page. This article is new, but is replacing another article that had a small group of highly valued contributors. We want those contributors to be listed as contributors to this new article. Each contributor will be asked to remove one bar in a separate PR, so they can then appear as a contributor to this new article.