Update power-bi-visualization-export-data.md
@tessahurr
Please review my copy edits. I'm not sure I copy edited the original version, the version you wrote, or the version Justin wrote. 
My understanding is that when you export, using either method, you get ALL the hierarchy data -- not just the data that was used to create the currently drilled version of the visual.  If my understanding is incorrect for one or both, please let me know.  Thanks.