Update the "disable connectivity" section
Removing the "Disable connectivity" section as we do not have an option now to disable the XMLA EP connectivity. The setting in the workload settings has been removed to accommodate the new XMLA RW features. Once the RW feature goes live, this section will be tailored accordingly to address the new UI settings of XMLA EP connection settings.