Changing attachment size limit per PG note
From: Guilherme Boreki <Guilherme.Boreki@microsoft.com> 
Sent: Monday, August 17, 2020 3:50 PM
I believe the limit might be 25Mb for the entire message - all content+metadata+attachment(s) also both attachments count toward the maximum (the image preview and the file attachment). We might limit in code that both attachments are not larger than 24Mb so we have room for the other stuff. So I'd probably update the docs for it since the customer would not be able to leverage full 25Mb anyway