Update power-bi-data-sources.md
Maggie had recently added a footnote for Paxata to the table of not being supported in PBI RS. This applies to all new data sources added to Desktop going forward. This ask is to please generalize that note. I’d put it at the top of the page to make it clear that the table is about the Power BI service and corresponding Desktop. we use the same wording as above the the article that list out the datasources supported by Report server

-------
cc: @davidiseminger