Update power-bi-browsers.md
We have dropped support for Internet Explorer (IE) as of August 17, 2021. We have not yet blocked IE usage since we're still seeing about 5% of users show up in IE to app.powerbi.com. But we are working with tenant admins and making plans to block IE  by 2022 and need to make sure support docs no longer show IE in the supported browser list.
-------
cc: @mihart