Update service-gateway-sso-kerberos-sap-bw-gx64krb.md
Per discussion with the product group, configuring both libraries on the same gateway service is an unsupported scenario. It’s not recommended to configure both libraries on the same gateway server as it’ll lead to a mix of libraries. If users want to use both libraries, they should fully separate the gateway server. For example, configure gx64krb for server A then sapcrypto for server B. Any failure for gx64krb is not supported.
So I think it's better to clarify this on official doc. For more information, please refer to this ICM.