Security Whitepaper: Client-side auth
Client-side authentication is in the process of going through the rings and is about to hit public production. This change updates the security whitepaper to describe the new flows that are going to happen with client-side authentication, and moves the server-side authentication details down to the Q&A section at the end (since the server-side authentication will still be used as a fallback in the short term if client-side authentication fails).