Update desktop-excel-stunning-report.md
The original guide is overall great, but sometimes still not clear enough, as sometimes they (the original authors) just mention a name of a random functionality which I didn't know where to look for, especially in case of graph styles, where there are only little icons. I sometimes had to guess where a functionality is. Moreover, it is already somewhat outdated, as some functionalities/features are now under a different name.