Update desktop-access-database-errors.md
Updating out of date and inaccurate information regarding supported use of the Access Database Engine. The Access 365 Runtime is the only redistributable installation that's currently within mainstream support. Additionally, Access 365 is the only version of Access that has changes included to work with the PowerBI Gateway service account. Lastly, we don't support installing multiple instances of the Access Database Engines using different bitnesses. Office only supports one bitness across all Office products on the machine. There should only be one installation on the machine. Either that's an Office suite or the standalone Access 365 Runtime. The only remaining exception to that is Office 2019, this either requires Access 2013 Runtime or switching to a newer Office install. Pointed to know issues article for this tope. Please reach out if there are any questions.