Link to correct version of Report builder
The link to Report Builder is not correct as it takes the user to  Power BI Report Builder's information which is the client for the Power BI service. and that's not the correct client.
To create paginated reports for Power BI Report Server the link should take you to Microsoft Report Builder.